Extract SorbetTypeVariables gem listener
