+++ lib/tapioca/compilers/symbol_table_compiler.rb
@@ -28,6 +28,7 @@ module Tapioca
         @bootstrap_symbols.each { |symbol| push_symbol(symbol) }
 
         @node_listeners = T.let([], T::Array[Gem::Listeners::Base])
+        @node_listeners << Gem::Listeners::SorbetTypeVariables.new(self)
         @node_listeners << Gem::Listeners::Mixins.new(self)
         @node_listeners << Gem::Listeners::DynamicMixins.new(self)
         @node_listeners << Gem::Listeners::SorbetHelpers.new(self)
@@ -209,7 +210,6 @@ module Tapioca
       sig { params(tree: RBI::Tree, name: String, constant: Module).void }
       def compile_body(tree, name, constant)
         # Compiling type variables must happen first to populate generic names
-        compile_type_variables(tree, constant)
         compile_methods(tree, name, constant)
       end
 
@@ -226,38 +226,6 @@ module Tapioca
 
           push_constant(symbol, subconstant)
         end
-      end
-
-      sig { params(tree: RBI::Tree, constant: Module).void }
-      def compile_type_variables(tree, constant)
-        compile_type_variable_declarations(tree, constant)
-
-        sclass = RBI::SingletonClass.new
-        compile_type_variable_declarations(sclass, singleton_class_of(constant))
-        tree << sclass if sclass.nodes.length > 1
-      end
-
-      sig { params(tree: RBI::Tree, constant: Module).void }
-      def compile_type_variable_declarations(tree, constant)
-        # Try to find the type variables defined on this constant, bail if we can't
-        type_variables = GenericTypeRegistry.lookup_type_variables(constant)
-        return unless type_variables
-
-        # Map each type variable to its string representation.
-        #
-        # Each entry of `type_variables` maps a Module to a String,
-        # and the order they are inserted into the hash is the order they should be
-        # defined in the source code.
-        type_variable_declarations = type_variables.map do |type_variable|
-          type_variable_name = type_variable.name
-          next unless type_variable_name
-
-          tree << RBI::TypeMember.new(type_variable_name, type_variable.serialize)
-        end
-
-        return if type_variable_declarations.empty?
-
-        tree << RBI::Extend.new("T::Generic")
       end
 
       sig { params(constant: Class).returns(T.nilable(String)) }
