+++ lib/tapioca/internal.rb
@@ -24,6 +24,7 @@ require "tapioca/compilers/gem/listeners/sorbet_enums"
 require "tapioca/compilers/gem/listeners/sorbet_helpers"
 require "tapioca/compilers/gem/listeners/sorbet_props"
 require "tapioca/compilers/gem/listeners/sorbet_required_ancestors"
+require "tapioca/compilers/gem/listeners/sorbet_type_variables"
 require "tapioca/compilers/gem/listeners/yard_doc"
 require "tapioca/compilers/requires_compiler"
 require "tapioca/compilers/symbol_table_compiler"
